backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyjwt>=2.8.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-anyio>=4.0.0
redis>=5.0.0
